+++ gsa/src/web/pages/reports/listpage.js
@@ -19,6 +19,8 @@
 
 import React from 'react';
 
+import {connect} from 'react-redux';
+
 import _ from 'gmp/locale';
 
 import Filter, {REPORTS_FILTER_FILTER} from 'gmp/models/filter';
@@ -26,7 +28,6 @@ import Filter, {REPORTS_FILTER_FILTER} from 'gmp/models/filter';
 import {isActive} from 'gmp/models/task';
 
 import {isDefined} from 'gmp/utils/identity';
-import {selectSaveId} from 'gmp/utils/id';
 
 import EntitiesPage from 'web/entities/page';
 import withEntitiesContainer from 'web/entities/withEntitiesContainer';
@@ -46,6 +47,11 @@ import {
   selector as entitiesSelector,
 } from 'web/store/entities/reports';
 
+import {
+  loadAllEntities as loadAllTasks,
+  selector as tasksSelector,
+} from 'web/store/entities/tasks';
+
 import {DEFAULT_RELOAD_INTERVAL_ACTIVE} from 'web/utils/constants';
 import compose from 'web/utils/compose';
 import PropTypes from 'web/utils/proptypes';
@@ -56,6 +62,8 @@ import ImportReportDialog from './importdialog';
 import ReportsTable from './table';
 
 import ReportsDashboard, {REPORTS_DASHBOARD_ID} from './dashboard';
+
+const CONTAINER_TASK_FILTER = Filter.fromString('target=""');
 
 const ToolBarIcons = ({onUploadReportClick}) => (
   <IconDivider>
@@ -91,7 +99,6 @@ class Page extends React.Component {
     this.openImportDialog = this.openImportDialog.bind(this);
     this.handleCloseImportDialog = this.handleCloseImportDialog.bind(this);
     this.openCreateTaskDialog = this.openCreateTaskDialog.bind(this);
-    this.loadTasks = this.loadTasks.bind(this);
   }
 
   componentWillReceiveProps(next) {
@@ -108,26 +115,16 @@ class Page extends React.Component {
     }
   }
 
-  loadTasks() {
-    const {gmp} = this.props;
-    return gmp.tasks.get().then(response => {
-      const {data: tasks} = response;
-      return tasks.filter(task => task.isContainer());
-    });
-  }
-
   openCreateTaskDialog() {
     this.setState({containerTaskDialogVisible: true});
   }
 
   openImportDialog(task_id) {
-    this.loadTasks().then(tasks =>
+    this.props.loadTasks().then(() => {
       this.setState({
-        tasks,
-        task_id: selectSaveId(tasks),
         importDialogVisible: true,
-      }),
-    );
+      });
+    });
   }
 
   closeImportDialog() {
@@ -152,16 +149,12 @@ class Page extends React.Component {
 
   handleCreateContainerTask(data) {
     const {gmp} = this.props;
-    let task_id;
-    return gmp.task
-      .createContainer(data)
-      .then(response => {
-        const {data: task} = response;
-        task_id = task.id;
-      })
-      .then(this.loadTasks)
-      .then(tasks => this.setState({tasks, task_id}))
-      .then(() => this.closeContainerTaskDialog());
+    return gmp.task.createContainer(data).then(response => {
+      const {data: task} = response;
+      this.props.loadTasks();
+      this.setState({task_id: task.id});
+      this.closeContainerTaskDialog();
+    });
   }
 
   handleCloseContainerTask() {
@@ -205,12 +198,11 @@ class Page extends React.Component {
   }
 
   render() {
-    const {filter, onFilterChanged, onInteraction} = this.props;
+    const {filter, onFilterChanged, onInteraction, tasks} = this.props;
     const {
       containerTaskDialogVisible,
       importDialogVisible,
       task_id,
-      tasks,
     } = this.state;
 
     return (
@@ -267,6 +259,8 @@ Page.propTypes = {
   filter: PropTypes.filter,
   gmp: PropTypes.gmp.isRequired,
   history: PropTypes.object.isRequired,
+  loadTasks: PropTypes.func.isRequired,
+  tasks: PropTypes.arrayOf(PropTypes.model),
   onChanged: PropTypes.func.isRequired,
   onError: PropTypes.func.isRequired,
   onFilterChanged: PropTypes.func.isRequired,
@@ -278,8 +272,23 @@ const reportsReloadInterval = ({entities = [], defaultReloadInterval}) =>
     ? DEFAULT_RELOAD_INTERVAL_ACTIVE
     : defaultReloadInterval;
 
+const mapStateToProps = rootState => {
+  const sel = tasksSelector(rootState);
+  return {
+    tasks: sel.getAllEntities(CONTAINER_TASK_FILTER),
+  };
+};
+
+const mapDispatchToProps = (dispatch, {gmp}) => ({
+  loadTasks: () => dispatch(loadAllTasks(gmp)(CONTAINER_TASK_FILTER)),
+});
+
 export default compose(
   withGmp,
+  connect(
+    mapStateToProps,
+    mapDispatchToProps,
+  ),
   withEntitiesContainer('report', {
     entitiesSelector,
     loadEntities,